Me-Skin2/Me-Skin Prediction/requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-multipart
jinja2